requirements.txt
bleak>=0.19.0
PyYAML>=6.0 